requirements.txt
tweepy==4.14.0
anthropic>=0.40.0,<1.0
openai>=1.40.0,<2.0
apscheduler==3.10.4
feedparser==6.0.11
requests==2.32.3
python-dotenv==1.0.1
Pillow>=10.0,<11.0
matplotlib>=3.7,<4.0
mplfinance>=0.12.0a0

# Testing
pytest==8.3.2
pytest-mock==3.14.0
